requirements/base.txt
Flask==0.10.1
Flask-RESTful==0.3.5
Flask-WTF==0.12
python-decouple==3.0
Flask-Redis==0.1.0
